Working on advanced task 102

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,16 +9,16 @@
 int main(void)
 {
 	int start;
-	int fibonacci = 1;
-	int last = 1;
-	int aux;
+	long int fibonacci = 1;
+	long int last = 1;
+	long int aux;
 
 	for (start = 0; start < 50; start++)
 	{
 		if (start < 49)
-			printf("%d, ", fibonacci);
+			printf("%ld, ", fibonacci);
 		else
-			printf("%d", fibonacci);
+			printf("%ld", fibonacci);
 
 		aux = fibonacci;
